requirements.txt
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
branca==0.8.2
cachetools==6.2.6
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
deprecation==2.1.0
duckdb==1.4.4
folium==0.20.0
geopandas==1.1.2
gitdb==4.0.12
GitPython==3.1.46
gotrue==2.12.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.27.2
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.16.0
numpy==2.4.2
packaging==26.0
pandas==2.3.3
pillow==12.1.1
pluggy==1.6.0
postgrest==0.18.0
protobuf==6.33.5
pyarrow==23.0.1
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.11.0
pyogrio==0.12.1
pyproj==3.7.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
realtime==2.28.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
scipy==1.17.1
shapely==2.1.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.3
storage3==0.9.0
streamlit==1.54.0
streamlit-folium==0.26.2
supabase==2.10.0
supafunc==0.7.0
tenacity==9.1.4
toml==0.10.2
tornado==6.5.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
websockets==15.0.1
xyzservices==2025.11.0
google-cloud-bigquery
google-cloud-bigquery-storage
pyarrow
pandas-gbq
